+++ src/Extension/UrlExtension.php
@@ -47,13 +47,17 @@ class UrlExtension implements ExtensionInterface
      */
     public function register(Engine $engine) : void
     {
-        $engine->registerFunction('url', [$this, 'generateUrl']);
-        $engine->registerFunction('serverurl', [$this, 'generateServerUrl']);
-        $engine->registerFunction('route', [$this, 'getRouteResult']);
+        $engine->registerFunction('url', $this->urlHelper);
+        $engine->registerFunction('serverurl', $this->serverUrlHelper);
+        $engine->registerFunction('route', [$this->urlHelper, 'getRouteResult']);
     }
 
     /**
      * Get the RouteResult instance of UrlHelper, if any.
+     *
+     * @deprecated since 2.2.0; to be removed in 3.0.0. This method was originally
+     *     used internally to back the route() Plates function; we now register
+     *     the UrlHelper::getRouteResult callback directly.
      */
     public function getRouteResult() : ?RouteResult
     {
@@ -68,6 +72,9 @@ class UrlExtension implements ExtensionInterface
      *     - reuse_result_params (bool): indicates if the current RouteResult
      *       parameters will be used, defaults to true
      * @return string
+     * @deprecated since 2.2.0; to be removed in 3.0.0. This method was originally
+     *     used internally to back the url() Plates function; we now register
+     *     UrlHelper instance directly, as it is callable.
      */
     public function generateUrl(
         string $routeName = null,
@@ -81,6 +88,10 @@ class UrlExtension implements ExtensionInterface
 
     /**
      * Generate a fully qualified URI, relative to $path.
+     *
+     * @deprecated since 2.2.0; to be removed in 3.0.0. This method was originally
+     *     used internally to back the serverurl() Plates function; we now register
+     *     the ServerUrl instance directly, as it is callable.
      */
     public function generateServerUrl(string $path = null) : string
     {